# openapi.yaml
openapi: 3.1.0
info:
  title: HTMLTrust Directory API
  description: |
    Reference implementation of the HTMLTrust trust directory API. Provides endpoints
    for author management, content signing and verification, claims, and a federated
    trust directory with reputation tracking.
  version: 1.0.0
  contact:
    name: HTMLTrust
    url: https://github.com/ArcadeLabsInc/htmltrust-server-reference
    email: jason@jason-grey.com
servers:
  - url: https://api.contentsigning.example.com/v1
    description: Production server
  - url: https://api.staging.contentsigning.example.com/v1
    description: Staging server
...
security:
  - {} # No default security
components:
  securitySchemes:
    AuthorApiKey:
      type: apiKey
      in: header
      name: X-AUTHOR-API-KEY
      description: API key linked to a specific author for author-specific operations

    GeneralApiKey:
      type: apiKey
      in: header
      name: X-API-KEY
      description: General API key for authenticated operations

    AdminApiKey:
      type: apiKey
      in: header
      name: X-ADMIN-API-KEY
      description: Admin-level API key for administrative operations
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
      example:
        code: "UNAUTHORIZED"
        message: "Invalid API key"
    Author:
      type: object
      required:
        - id
        - name
        - keyType
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the author
        name:
          type: string
          description: Name of the author
        description:
          type: string
          description: Description of the author
        url:
          type: string
          format: uri
          description: URL associated with the author
        keyType:
          type: string
          enum: [HUMAN, AI, HUMAN_AI_MIX, ORGANIZATION]
          description: Type of the author key
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        name: "John Doe"
        description: "Technology writer and researcher"
        url: "https://johndoe.example.com"
        keyType: "HUMAN"
        createdAt: "2023-01-01T12:00:00Z"
        updatedAt: "2023-01-01T12:00:00Z"
...
    ContentSignature:
      type: object
      required:
        - contentHash
        - domain
        - authorId
        - signature
        - claims
      properties:
        contentHash:
          type: string
          description: Hash of the normalized content
        domain:
          type: string
          description: Domain associated with the content
        authorId:
          type: string
          format: uuid
          description: ID of the author who signed the content
        signature:
          type: string
          description: Cryptographic signature binding content, hash, domain, and author key
        claims:
          type: object
          description: Claims about the content
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
      example:
        contentHash: "sha256:a591a6d40bf420404a011733cfb7b190d62c65bf0bcda32b57b277d9ad9f146e"
        domain: "example.com"
        authorId: "123e4567-e89b-12d3-a456-426614174000"
        signature: "MEUCIQD7y5SxmQJ9f0lE9B0BwqIJKKdL5fZMNQOiPnKWUJfmrgIgEbHtPwDxM9xGbCZzW9k2R9jFxwJZQQlPfhgj+0YP7vQ="
        claims:
          ContentType: "Article"
          License: "CC-BY-4.0"
          AIAssistance: "Minimal"
        createdAt: "2023-01-01T12:00:00Z"
...
paths:
...
  /content/sign:
    post:
      tags:
        - Content
      summary: Sign content
      description: Signs content with the author's private key
      operationId: signContent
      security:
        - AuthorApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contentHash
                - domain
                - claims
              properties:
                contentHash:
                  type: string
                  description: Hash of the normalized content
                domain:
                  type: string
                  description: Domain associated with the content
                claims:
                  type: object
                  description: Claims about the content
                  additionalProperties: true
      responses:
        "201":
          description: Content signed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentSignature"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /content/verify:
    post:
      tags:
        - Content
      summary: Verify content signature
      description: Verifies the signature of content
      operationId: verifyContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contentHash
                - domain
                - authorId
                - signature
              properties:
                contentHash:
                  type: string
                  description: Hash of the normalized content
                domain:
                  type: string
                  description: Domain associated with the content
                authorId:
                  type: string
                  format: uuid
                  description: ID of the author who signed the content
                signature:
                  type: string
                  description: Cryptographic signature to verify
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                required:
                  - valid
                properties:
                  valid:
                    type: boolean
                    description: Whether the signature is valid
                  author:
                    $ref: "#/components/schemas/Author"
                  claims:
                    type: object
                    description: Claims included in the signature
                    additionalProperties: true
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...